Use ical4android that removes ClassLoader checks; refactor SyncDispatcher
