+++ app/src/main/kotlin/at/bitfire/davdroid/di/CoroutineDispatchersModule.kt
@@ -11,6 +11,7 @@ import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import javax.inject.Qualifier
+import javax.inject.Singleton
 
 @Retention(AnnotationRetention.RUNTIME)
 @Qualifier
@@ -23,6 +24,10 @@ annotation class IoDispatcher
 @Retention(AnnotationRetention.RUNTIME)
 @Qualifier
 annotation class MainDispatcher
+
+@Retention(AnnotationRetention.RUNTIME)
+@Qualifier
+annotation class SyncDispatcher
 
 @Module
 @InstallIn(SingletonComponent::class)
@@ -39,5 +44,11 @@ class CoroutineDispatchersModule {
     @Provides
     @MainDispatcher
     fun mainDispatcher(): CoroutineDispatcher = Dispatchers.Main
+
+    @Provides
+    @Singleton
+    @SyncDispatcher
+    fun syncDispatcher(): CoroutineDispatcher =
+        Dispatchers.Default.limitedParallelism(Runtime.getRuntime().availableProcessors())
 
 }
