+++ app/src/main/kotlin/at/bitfire/davdroid/resource/LocalEvent.kt
@@ -12,21 +12,23 @@ import android.provider.CalendarContract
 import android.provider.CalendarContract.Events
 import androidx.core.content.contentValuesOf
 import at.bitfire.davdroid.BuildConfig
+import at.bitfire.davdroid.resource.LocalEvent.Companion.numInstances
 import at.bitfire.ical4android.AndroidCalendar
 import at.bitfire.ical4android.AndroidEvent
 import at.bitfire.ical4android.AndroidEventFactory
 import at.bitfire.ical4android.BatchOperation
 import at.bitfire.ical4android.Event
 import at.bitfire.ical4android.ICalendar
-import at.bitfire.ical4android.Ical4Android
+import at.bitfire.ical4android.ical4jVersion
 import at.bitfire.ical4android.util.MiscUtils.asSyncAdapter
 import net.fortuna.ical4j.model.property.ProdId
 import java.util.UUID
+
 class LocalEvent: AndroidEvent, LocalResource<Event> {
 
     companion object {
         init {
-            ICalendar.prodId = ProdId("DAVx5/${BuildConfig.VERSION_NAME} ical4j/" + Ical4Android.ical4jVersion)
+            ICalendar.prodId = ProdId("DAVx5/${BuildConfig.VERSION_NAME} ical4j/" + ical4jVersion)
         }
 
         const val COLUMN_ETAG = Events.SYNC_DATA1
