+++ app/src/main/kotlin/at/bitfire/davdroid/sync/SyncManager.kt
@@ -41,7 +41,6 @@ import at.bitfire.davdroid.sync.SyncManager.Companion.DELAY_UNTIL_MAX
 import at.bitfire.davdroid.sync.SyncManager.Companion.DELAY_UNTIL_MIN
 import at.bitfire.davdroid.sync.account.InvalidAccountException
 import at.bitfire.ical4android.CalendarStorageException
-import at.bitfire.ical4android.Ical4Android
 import at.bitfire.vcard4android.ContactsStorageException
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.coroutineScope
@@ -151,11 +150,6 @@ abstract class SyncManager<ResourceType: LocalResource<*>, out CollectionType: L
     @Inject
     lateinit var syncNotificationManagerFactory: SyncNotificationManager.Factory
 
-
-    init {
-        // required for ServiceLoader -> ical4j -> ical4android
-        Ical4Android.checkThreadContextClassLoader()
-    }
 
     protected lateinit var davCollection: RemoteType
