+++ app/src/main/kotlin/at/bitfire/davdroid/sync/worker/OneTimeSyncWorker.kt
@@ -12,11 +12,12 @@ import androidx.work.ForegroundInfo
 import androidx.work.WorkManager
 import androidx.work.WorkerParameters
 import at.bitfire.davdroid.R
+import at.bitfire.davdroid.di.SyncDispatcher
 import at.bitfire.davdroid.sync.SyncDataType
-import at.bitfire.davdroid.sync.SyncDispatcher
 import at.bitfire.davdroid.ui.NotificationRegistry
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
+import kotlinx.coroutines.CoroutineDispatcher
 
 /**
  * One-time sync worker.
@@ -29,8 +30,8 @@ import dagger.assisted.AssistedInject
 class OneTimeSyncWorker @AssistedInject constructor(
     @Assisted appContext: Context,
     @Assisted workerParams: WorkerParameters,
-    syncDispatcher: SyncDispatcher
-) : BaseSyncWorker(appContext, workerParams, syncDispatcher.dispatcher) {
+    @SyncDispatcher syncDispatcher: CoroutineDispatcher
+) : BaseSyncWorker(appContext, workerParams, syncDispatcher) {
 
     /**
      * Used by WorkManager to show a foreground service notification for expedited jobs on Android <12.
