+++ app/src/main/kotlin/at/bitfire/davdroid/sync/worker/PeriodicSyncWorker.kt
@@ -10,11 +10,12 @@ import androidx.annotation.VisibleForTesting
 import androidx.hilt.work.HiltWorker
 import androidx.work.WorkManager
 import androidx.work.WorkerParameters
+import at.bitfire.davdroid.di.SyncDispatcher
 import at.bitfire.davdroid.sync.SyncDataType
-import at.bitfire.davdroid.sync.SyncDispatcher
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
+import kotlinx.coroutines.CoroutineDispatcher
 
 /**
  * Handles scheduled sync requests.
@@ -36,8 +37,8 @@ import dagger.assisted.AssistedInject
 class PeriodicSyncWorker @AssistedInject constructor(
     @Assisted appContext: Context,
     @Assisted workerParams: WorkerParameters,
-    syncDispatcher: SyncDispatcher
-) : BaseSyncWorker(appContext, workerParams, syncDispatcher.dispatcher) {
+    @SyncDispatcher syncDispatcher: CoroutineDispatcher
+) : BaseSyncWorker(appContext, workerParams, syncDispatcher) {
 
     @AssistedFactory
     @VisibleForTesting
