+++ gradle/libs.versions.toml
@@ -20,7 +20,7 @@ androidx-test-junit = "1.2.1"
 androidx-work = "2.10.1"
 bitfire-cert4android = "b67ba86d31"
 bitfire-dav4jvm = "7d0c4f6fb9"
-bitfire-ical4android = "ef907cf6b4"
+bitfire-ical4android = "21622fef29"
 bitfire-vcard4android = "59eb998f29"
 compose-accompanist = "0.37.2"
 compose-bom = "2025.04.01"
